优化计算嫌疑度算法，让计算更加精确         20250729

+++ cigarette-system/src/main/java/com/ruoyi/system/mapper/ToVehicleRealTimMonitoringMapper.java
@@ -15,5 +15,5 @@ public interface ToVehicleRealTimMonitoringMapper {
 
     List<ToVehicleRealTimMonitoring> selectAfterId(@Param("lastId") Long lastId, @Param("limit") int limit);
 
-    int updateLevel(@Param("id") Long id, @Param("level") Integer level);
+    int updateLevel(@Param("id") Long id, @Param("level") Double level);
 }
